requirements.txt
Jinja2
PyYAML>=5.3,<6.0
